chore: move and unify SERVICE_NAME
