+++ src/Components/ServiceScene/ActionBarScene.tsx
@@ -4,8 +4,7 @@ import { getExplorationFor } from '../../services/scenes';
 import { getDrilldownSlug, getDrilldownValueSlug, PageSlugs, ValueSlugs } from '../../services/routing';
 import { GoToExploreButton } from './GoToExploreButton';
 import { reportAppInteraction, USER_EVENTS_ACTIONS, USER_EVENTS_PAGES } from '../../services/analytics';
-import { ALL_VARIABLE_VALUE, getLabelsVariable } from '../../services/variables';
-import { SERVICE_NAME } from '../ServiceSelectionScene/ServiceSelectionScene';
+import { ALL_VARIABLE_VALUE, getLabelsVariable, SERVICE_NAME } from '../../services/variables';
 import { navigateToDrilldownPage, navigateToIndex } from '../../services/navigate';
 import React from 'react';
 import { ServiceScene, ServiceSceneState } from './ServiceScene';
