+++ src/Components/ServiceScene/LogsPanelScene.tsx
@@ -15,9 +15,8 @@ import { LogsListScene } from './LogsListScene';
 import { LoadingPlaceholder } from '@grafana/ui';
 import { addToFilters, FilterType } from './Breakdowns/AddToFiltersButton';
 import { getLabelTypeFromFrame, LabelType } from '../../services/fields';
-import { getAdHocFiltersVariable, VAR_FIELDS, VAR_LABELS, VAR_LEVELS } from '../../services/variables';
+import { getAdHocFiltersVariable, SERVICE_NAME, VAR_FIELDS, VAR_LABELS, VAR_LEVELS } from '../../services/variables';
 import { reportAppInteraction, USER_EVENTS_ACTIONS, USER_EVENTS_PAGES } from '../../services/analytics';
-import { SERVICE_NAME } from '../ServiceSelectionScene/ServiceSelectionScene';
 
 interface LogsPanelSceneState extends SceneObjectState {
   body?: VizPanel;
