+++ src/Components/ServiceScene/ServiceScene.tsx
@@ -24,6 +24,7 @@ import {
   getLabelsVariable,
   getLevelsVariable,
   LOG_STREAM_SELECTOR_EXPR,
+  SERVICE_NAME,
   VAR_DATASOURCE,
   VAR_FIELDS,
   VAR_LABELS,
@@ -31,7 +32,6 @@ import {
   VAR_LEVELS,
   VAR_PATTERNS,
 } from 'services/variables';
-import { SERVICE_NAME } from 'Components/ServiceSelectionScene/ServiceSelectionScene';
 import { getMetadataService } from '../../services/metadata';
 import { navigateToIndex } from '../../services/navigate';
 import { areArraysEqual } from '../../services/comparison';
